Introduce SqlValidateStorageTypeCombinationFeature to do the validation the core should be doing in the right place
