+++ src/SqlPersistence/SynchronizedStorage/SqlStorageSessionFeature.cs
@@ -1,17 +1,13 @@
-﻿using System;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using NServiceBus;
 using NServiceBus.Features;
 using NServiceBus.Persistence;
-using NServiceBus.Settings;
 
 class SqlStorageSessionFeature : Feature
 {
     protected override void Setup(FeatureConfigurationContext context)
     {
         var settings = context.Settings;
-        ValidateSagaOutboxCombo(settings);
-
         var sqlDialect = settings.GetSqlDialect();
         var services = context.Services;
         var connectionManager = settings.GetConnectionBuilder<StorageType.Sagas>();
@@ -24,24 +20,5 @@ class SqlStorageSessionFeature : Feature
 
         services.AddTransient(_ => sessionHolder.Current);
         context.Pipeline.Register(new CurrentSessionBehavior(sessionHolder), "Manages the lifecycle of the current session holder.");
-    }
-
-    // This check should normally be handled by Core but unfortunately there is a bug that prevents the check from executing
-    // https://github.com/Particular/NServiceBus/issues/6378
-    static void ValidateSagaOutboxCombo(IReadOnlySettings settings)
-    {
-        var isOutboxEnabled = settings.IsFeatureActive(typeof(Outbox));
-        var isSagasEnabled = settings.IsFeatureActive(typeof(Sagas));
-        if (!isOutboxEnabled || !isSagasEnabled)
-        {
-            return;
-        }
-        var isSagasEnabledForSqlPersistence = settings.IsFeatureActive(typeof(SqlSagaFeature));
-        var isOutboxEnabledForSqlPersistence = settings.IsFeatureActive(typeof(SqlOutboxFeature));
-        if (isSagasEnabledForSqlPersistence && isOutboxEnabledForSqlPersistence)
-        {
-            return;
-        }
-        throw new Exception("Sql Persistence must be enabled for either both Sagas and Outbox, or neither.");
     }
 }
